+++ src/heap.h
@@ -8,7 +8,7 @@ typedef struct edge_s
 {
   int vertix1 ;
   int vertix2 ;
-  long double weight ;
+  float weight ;
 } edge_t ;
 
 
@@ -35,7 +35,7 @@ typedef struct heap_s
    @param poids Le poids du trajet.
    @return Un pointer vers une arête paramétré.
  */
-edge_t edge_create(int n1, int n2, long double poids);
+edge_t edge_create(int n1, int n2, float poids);
 
 
 /**
